Private notes and questions Fixed audio uploading/playing for notes (still slow)

+++ src/main/java/rs/luka/android/studygroup/network/Network.java
@@ -268,6 +268,7 @@ public class Network {
         @Override
         protected File getData(InputStream stream) throws IOException {
             if(saveTo != null) {
+                if(!saveTo.exists() && !saveTo.createNewFile()) throw new IOException("Cannot create new file");
                 FileOutputStream out  = new FileOutputStream(saveTo);
                 byte[]           buff = new byte[BUFFER_SIZE];
                 int              readBytes;
